Enable source-build with arcade
This enables 'source-build', which makes it easier to build the entire
shipping .NET SDK from source.

This is the first and second step of arcade-powered-source-build:
https://github.com/dotnet/source-build/blob/master/Documentation/planning/arcade-powered-source-build/README.md

See dotnet/sourcelink#692 for a similar PR, that this is based on.

Aside from arcade-specific changes, it also makes all `.sh` files
executable to make this build under Linux/macOS.

The rename from LICENSE to LICENSE.txt is a workaround for
https://github.com/NuGet/Home/issues/7601.