Make 'using' declarations global
One of C# 10's new features is the 'global using'. This enables a single
'using' to be applied to all source code files within a project. Given
that these are test projects, it is likely that the majority of source
code files within them will require inclusion of a reference to the test
framework. Therefore it seems like these particular usings are excellent
candidates to be made global.